added authoratization to courses endpoints

+++ api/courses.js
@@ -17,12 +17,17 @@ const csv = require("csv-parser");
 const csvWriter = require("csv-writer");
 const { Transform } = require("stream");
 const { ObjectId } = require("mongodb");
+const { verifyUser } = require("./auth");
 
 const router = Router();
 
-router.post("/", async (req, res, next) => {
+router.post("/", verifyUser, async (req, res, next) => {
   console.log(" -- req.body: ", req.body);
   if (validateAgainstSchema(req.body, CourseSchema)) {
+    if (req.user.role !== "admin") {
+      res.status(403).json({ error: "Unauthorized!" });
+      return;
+    }
     try {
       console.log("  -- Create a new course");
       const id = await insertNewCourse(req.body);
@@ -75,10 +80,19 @@ router.get("/:id", async (req, res, next) => {
   }
 });
 
-router.patch("/:id", async (req, res, next) => {
+router.patch("/:id", verifyUser, async (req, res, next) => {
   try {
     console.log(`  -- Update data about a specific course: ${req.params.id}`);
     if (validateAgainstSchema(req.body, CourseSchema)) {
+      const course = getCourseById(req.params.id);
+      const instructorId = course.instructorId;
+      if (
+        req.user.role !== "admin" &&
+        !(req.user.role === "instructor" && req.user.userId === instructorId)
+      ) {
+        res.status(403).json({ error: "Unauthorized!" });
+        return;
+      }
       const updateSuccessful = await updateCourseById(req.params.id, req.body);
       //   console.log("  -- updateSuccessful: ", updateSuccessful);
       if (updateSuccessful) {
@@ -96,9 +110,13 @@ router.patch("/:id", async (req, res, next) => {
   }
 });
 
-router.delete("/:id", async (req, res, next) => {
+router.delete("/:id", verifyUser, async (req, res, next) => {
   try {
     console.log(`  -- Remove a specific course: ${req.params.id}`);
+    if (req.user.role !== "admin") {
+      res.status(403).json({ error: "Unauthorized!" });
+      return;
+    }
     const deleteSuccessful = await deleteCourseById(req.params.id);
     // console.log("deleteSuc: ", deleteSuccessful);
     if (deleteSuccessful) {
@@ -111,8 +129,18 @@ router.delete("/:id", async (req, res, next) => {
   }
 });
 
-router.post("/:id/students", async (req, res, next) => {
+router.post("/:id/students", verifyUser, async (req, res, next) => {
   if (req.body && (req.body.add || req.body.remove)) {
+    const course = getCourseById(req.params.id);
+    const instructorId = course.instructorId;
+    console.log(" -- role:", req.user.role);
+    if (
+      req.user.role !== "admin" &&
+      !(req.user.role === "instructor" && req.user.userId === instructorId)
+    ) {
+      res.status(403).json({ error: "Unauthorized!" });
+      return;
+    }
     try {
       console.log(
         `  -- Update enrollment for a specific course: ${req.params.id}`
@@ -149,7 +177,16 @@ router.post("/:id/students", async (req, res, next) => {
   }
 });
 
-router.get("/:id/students", async (req, res, next) => {
+router.get("/:id/students", verifyUser, async (req, res, next) => {
+  const course = getCourseById(req.params.id);
+  const instructorId = course.instructorId;
+  if (
+    req.user.role !== "admin" &&
+    !(req.user.role === "instructor") & (req.user.userId === instructorId)
+  ) {
+    res.status(403).json({ error: "Unauthorized!" });
+    return;
+  }
   try {
     console.log(
       `  -- Fetch a list of the students enrolled in a specific course: ${req.params.id}`
